pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "f1-strategy-lab"
version = "0.1.0"
description = "End-to-end ML + CV race strategy system for Formula 1"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Sumukh" }]
dependencies = [
  "fastf1>=3.5.0",
  "numpy>=1.26.0",
  "pandas>=2.2.0",
  "scikit-learn>=1.5.0",
  "scipy>=1.13.0",
  "opencv-python>=4.10.0",
  "requests>=2.31.0",
  "pyyaml>=6.0.1",
  "joblib>=1.4.0",
  "typer>=0.12.0",
  "rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.5.0",
]

[project.scripts]
f1lab-run = "f1_strategy_lab.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py310"

